Port Preserved[Module|Function]HashAnalysis to Analysis
